requirements.txt
faiss_cpu==1.7.3
langchain==0.0.136
pydantic==1.10.7
python-dotenv==1.0.0
llama_cpp_python==0.1.33
